remove old bbconf
