+++ bbconf/bbconf.py
@@ -1,95 +1,120 @@
-import psycopg2
-from psycopg2 import sql
-from psycopg2.extras import DictCursor, Json
-from psycopg2.extensions import connection
-
+import os
 from logging import getLogger
-from contextlib import contextmanager
-from collections.abc import Mapping
-from re import findall
+from psycopg2 import sql
 
 import yacman
-import os
+import pipestat
+from pipestat.const import *
 
-from attmap import PathExAttMap as PXAM
 
-from .const import *
 from .exceptions import *
+from .const import *
+from .helpers import *
 
 _LOGGER = getLogger(PKG_NAME)
 
-__all__ = ["BedBaseConf", "get_bedbase_cfg"]
 
-
-class LoggingCursor(psycopg2.extras.DictCursor):
-    """
-    Logging db cursor
-    """
-    def execute(self, query, vars=None):
-        """
-        Execute a database operation (query or command) and issue a debug
-        and info level log messages
-
-        :param query:
-        :param vars:
-        :return:
-        """
-        _LOGGER.debug(f"Executing query: {self.mogrify(query, vars)}")
-        try:
-            super(LoggingCursor, self).execute(query=query, vars=vars)
-        except Exception as e:
-            _LOGGER.error(f"{e.__class__.__name__}: {e}")
-            raise
-        else:
-            _LOGGER.debug(f"Executed query: {self.query}")
-
-
-class BedBaseConf(yacman.YacAttMap):
+class BedBaseConf(dict):
     """
     This class provides is an in-memory representation of the configuration
     file for the *BEDBASE* project. Additionally it implements multiple
     convenience methods for interacting with the database backend,
     i.e. [PostgreSQL](https://www.postgresql.org/)
     """
-    def __init__(self, filepath):
+    def __init__(self, config_path=None, database_only=False):
         """
-        Create the config instance with a filepath
+        Initialize the object
 
-        :param str filepath: a path to the YAML file to read
+        :param str config_path: path to the bedbase configuration file
+        :param bool database_only: whether the database managers should not
+            keep an in-memory copy of the data in the database
         """
-
         def _raise_missing_key(key):
             raise MissingConfigDataError("Config lacks '{}' key".format(key))
 
-        super(BedBaseConf, self).__init__(filepath=filepath)
+        super(BedBaseConf, self).__init__()
 
-        if CFG_PATH_KEY not in self:
+        cfg_path = get_bedbase_cfg(config_path)
+        self[CONFIG_KEY] = yacman.YacAttMap(filepath=cfg_path)
+        if CFG_PATH_KEY not in self[CONFIG_KEY]:
             _raise_missing_key(CFG_PATH_KEY)
-        if not self[CFG_PATH_KEY]:
+        if not self[CONFIG_KEY][CFG_PATH_KEY]:
             # if there's nothing under path key (None)
-            self[CFG_PATH_KEY] = PXAM()
-
-        if CFG_PIPELINE_OUT_PTH_KEY not in self[CFG_PATH_KEY]:
+            self[CONFIG_KEY][CFG_PATH_KEY] = {}
+        if CFG_PIPELINE_OUT_PTH_KEY not in self[CONFIG_KEY][CFG_PATH_KEY]:
             _raise_missing_key(CFG_PIPELINE_OUT_PTH_KEY)
 
-        if CFG_BEDSTAT_DIR_KEY not in self[CFG_PATH_KEY]:
+        if CFG_BEDSTAT_DIR_KEY not in self[CONFIG_KEY][CFG_PATH_KEY]:
             _raise_missing_key(CFG_BEDSTAT_DIR_KEY)
-            
-        if CFG_BEDBUNCHER_DIR_KEY not in self[CFG_PATH_KEY]:
-            _raise_missing_key(CFG_BEDBUNCHER_DIR_KEY)
-        
-        for section, mapping in DEFAULT_SECTION_VALUES.items():
-            if section not in self:
-                self[section] = PXAM()
-            for key, default in mapping.items():
-                if key not in self[section]:
-                    _LOGGER.debug("Config lacks '{}.{}' key. Setting to: {}".
-                                  format(section, key, default))
-                    self[section][key] = default
 
-    def _excl_from_repr(self, k, cls):
-        return k in \
-               [attr for attr in self.to_dict().keys() if attr.startswith("_")]
+        if CFG_BEDBUNCHER_DIR_KEY not in self[CONFIG_KEY][CFG_PATH_KEY]:
+            _raise_missing_key(CFG_BEDBUNCHER_DIR_KEY)
+
+        for section, mapping in DEFAULT_SECTION_VALUES.items():
+            if section not in self[CONFIG_KEY]:
+                self[CONFIG_KEY][section] = {}
+            for key, default in mapping.items():
+                if key not in self[CONFIG_KEY][section]:
+                    _LOGGER.debug(f"Config lacks '{section}.{key}' key. "
+                                  f"Setting to: {default}")
+                    self[CONFIG_KEY][section][key] = default
+
+        self[PIPESTATS_KEY] = {}
+        self[PIPESTATS_KEY][BED_TABLE] = pipestat.PipestatManager(
+            name=BED_TABLE,
+            config=self.config,
+            schema_path=BED_TABLE_SCHEMA,
+            database_only=database_only
+        )
+        self[PIPESTATS_KEY][BEDSET_TABLE] = pipestat.PipestatManager(
+            name=BEDSET_TABLE,
+            config=self.config,
+            schema_path=BEDSET_TABLE_SCHEMA,
+            database_only=database_only
+        )
+
+    def __str__(self):
+        """
+        Generate string representation of the object
+
+        :return str: string representation of the object
+        """
+        from textwrap import indent
+        res = f"{self.__class__.__name__}\n"
+        res += f"{BED_TABLE}:\n"
+        res += f"{indent(str(self.bed), '  ')}"
+        res += f"\n{BEDSET_TABLE}:\n"
+        res += f"{indent(str(self.bedset), '  ')}"
+        res += f"\nconfig:\n"
+        res += f"{indent(str(self.config), '  ')}"
+        return res
+
+    @property
+    def config(self):
+        """
+        Config used to initialize the object
+
+        :return yacman.YacAttMap: bedbase configuration file contents
+        """
+        return self[CONFIG_KEY]
+
+    @property
+    def bed(self):
+        """
+        PipestatManager of the bedfiles table
+
+        :return pipestat.PipestatManager: manager of the bedfiles table
+        """
+        return self[PIPESTATS_KEY][BED_TABLE]
+
+    @property
+    def bedset(self):
+        """
+        PipestatManager of the bedsets table
+
+        :return pipestat.PipestatManager: manager of the bedsets table
+        """
+        return self[PIPESTATS_KEY][BEDSET_TABLE]
 
     def _get_output_path(self, table_name, remote=False):
         """
@@ -102,12 +127,12 @@ class BedBaseConf(yacman.YacAttMap):
         """
         dir_key = CFG_BEDBUNCHER_DIR_KEY if table_name == BEDSET_TABLE \
             else CFG_BEDSTAT_DIR_KEY
-        base = self[CFG_PATH_KEY][CFG_REMOTE_URL_BASE_KEY] if remote \
-            else self[CFG_PATH_KEY][CFG_PIPELINE_OUT_PTH_KEY]
+        base = self.config[CFG_PATH_KEY][CFG_REMOTE_URL_BASE_KEY] if remote \
+            else self.config[CFG_PATH_KEY][CFG_PIPELINE_OUT_PTH_KEY]
         if remote and not base:
             raise MissingConfigDataError(
                 f"{CFG_REMOTE_URL_BASE_KEY} key value is invalid: {base}")
-        return os.path.join(base, self[CFG_PATH_KEY][dir_key])
+        return os.path.join(base, self.config[CFG_PATH_KEY][dir_key])
 
     def get_bedbuncher_output_path(self, remote=False):
         """
@@ -127,303 +152,7 @@ class BedBaseConf(yacman.YacAttMap):
         """
         return self._get_output_path(table_name=BED_TABLE, remote=remote)
 
-    def check_connection(self):
-        """
-        Check whether a PostgreSQL connection has been established
-
-        :return bool: whether the connection has been established
-        """
-        if hasattr(self, PG_CLIENT_KEY) and \
-                isinstance(getattr(self, PG_CLIENT_KEY), connection):
-            return True
-        return False
-
-    def establish_postgres_connection(self, suppress=False):
-        """
-        Establish PostgreSQL connection using the config data
-
-        :param bool suppress: whether to suppress any connection errors
-        :return bool: whether the connection has been established successfully
-        """
-        if self.check_connection():
-            raise BedBaseConnectionError(
-                f"The connection is already established: "
-                f"{self[PG_CLIENT_KEY].info.host}")
-        db = self[CFG_DATABASE_KEY]
-        try:
-            self[PG_CLIENT_KEY] = psycopg2.connect(
-                dbname=db[CFG_NAME_KEY],
-                user=db[CFG_USER_KEY],
-                password=db[CFG_PASSWORD_KEY],
-                host=self[CFG_DATABASE_KEY][CFG_HOST_KEY],
-                port=db[CFG_PORT_KEY]
-            )
-        except psycopg2.Error as e:
-            _LOGGER.error(f"Could not connect to: "
-                          f"{self[CFG_DATABASE_KEY][CFG_HOST_KEY]}")
-            _LOGGER.info(f"Caught error: {e}")
-            if suppress:
-                return False
-            raise
-        else:
-            _LOGGER.debug(f"Established connection with PostgreSQL: "
-                          f"{self[CFG_DATABASE_KEY][CFG_HOST_KEY]}")
-            return True
-
-    def close_postgres_connection(self):
-        """
-        Close connection and remove client bound
-        """
-        if not self.check_connection():
-            raise BedBaseConnectionError(
-                f"The connection has not been established: "
-                f"{self[CFG_DATABASE_KEY][CFG_HOST_KEY]}")
-        self[PG_CLIENT_KEY].close()
-        del self[PG_CLIENT_KEY]
-        _LOGGER.debug(f"Closed connection with PostgreSQL: "
-                      f"{self[CFG_DATABASE_KEY][CFG_HOST_KEY]}")
-
-    @property
-    @contextmanager
-    def db_cursor(self):
-        """
-        Establish connection and get a PostgreSQL database cursor,
-        commit and close the connection afterwards
-
-        :return DictCursor: Database cursor object
-        """
-        try:
-            if not self.check_connection():
-                self.establish_postgres_connection()
-            conn = self[PG_CLIENT_KEY]
-            with conn as c, c.cursor(cursor_factory=LoggingCursor) as cur:
-                yield cur
-        except:
-            raise
-        finally:
-            self.close_postgres_connection()
-
-    def _select_all(self, table_name):
-        """
-        Get all the contents from the selected table.
-        Convenience method for table exploration
-
-        :param str table_name: name of the table to list contents for
-        :return list[psycopg2.extras.DictRow]: all table contents
-        """
-        with self.db_cursor as cur:
-            cur.execute(
-                sql.SQL("SELECT * FROM {}").format(sql.Identifier(table_name)))
-            result = cur.fetchall()
-        return result
-
-    def select(self, table_name, columns=None, condition=None, condition_val=None):
-        """
-        Get all the contents from the selected table, possibly restricted by
-        the provided condition.
-
-        :param str table_name: name of the table to list contents for
-        :param str | list[str] columns: columns to select
-        :param str condition: condition to restrict the results
-            with, will be appended to the end of the SELECT statement and
-            safely populated with 'condition_val',
-            for example: `"id=%s"`
-        :param list condition_val: values to fill the placeholder
-            in 'condition' with
-        :return list[psycopg2.extras.DictRow]: all table contents
-        """
-        condition, condition_val = \
-            _preprocess_condition_pair(condition, condition_val)
-        if not columns:
-            columns = sql.SQL("*")
-        else:
-            columns = sql.SQL(',').join(
-                [sql.Identifier(x) for x in _mk_list_of_str(columns)])
-        statement = sql.SQL("SELECT {} FROM {}").format(
-            columns, sql.Identifier(table_name))
-        if condition:
-            statement += sql.SQL(" WHERE ")
-            statement += condition
-        with self.db_cursor as cur:
-            cur.execute(query=statement, vars=condition_val)
-            result = cur.fetchall()
-        return result
-
-    def _insert(self, table_name, values, skip_id_return=False):
-        """
-
-        :param str table_name: name of the table to insert the values to
-        :param dict values: a mapping of pairs of table column names and
-            respective values to be inserted to the database
-        :return int: id of the row just inserted
-        """
-        statement = sql.SQL("INSERT INTO {table_name} ({columns}) VALUES ({placeholder})").format(
-            placeholder=sql.SQL(','.join(['%s'] * len(values))),
-            columns=sql.SQL(',').join([sql.Identifier(v) for v in values.keys()]),
-            table_name=sql.Identifier(table_name))
-        if not skip_id_return:
-            statement += sql.SQL(" RETURNING id")
-        # convert mappings to JSON for postgres
-        values = tuple([Json(v) if isinstance(v, Mapping) else v
-                        for v in list(values.values())])
-        with self.db_cursor as cur:
-            cur.execute(statement, values)
-            if not skip_id_return:
-                return cur.fetchone()[0]
-
-    def insert_bedfile_data(self, values):
-        """
-
-        :param dict values: a mapping of pairs of table column names and
-            respective values to bne inserted to the database
-        :return int: id of the row just inserted
-        """
-        return self._insert(table_name=BED_TABLE, values=values)
-
-    def insert_bedset_data(self, values):
-        """
-
-        :param dict values: a mapping of pairs of table column names and
-            respective values to bne inserted to the database
-        :return int: id of the row just inserted
-        """
-        return self._insert(table_name=BEDSET_TABLE, values=values)
-
-    def insert_bedset_bedfiles_data(self, values):
-        """
-
-        :param dict values: a mapping of pairs of table column names and
-            respective values to bne inserted to the database
-        """
-        return self._insert(table_name=REL_TABLE, values=values, skip_id_return=True)
-
-    def _count_rows(self, table_name):
-        """
-        Count rows in a selected table
-
-        :param str table_name: table to count rows for
-        :return int: number of rows in the selected table
-        """
-        with self.db_cursor as cur:
-            statement = sql.SQL("SELECT COUNT(*) FROM {}").format(
-                sql.Identifier(table_name))
-            cur.execute(statement)
-            return cur.fetchall()[0][0]
-
-    def count_bedfiles(self):
-        """
-        Count rows in the bedfiles table
-
-        :return int: number of rows in the bedfiles table
-        """
-        return self._count_rows(table_name=BED_TABLE)
-
-    def count_bedsets(self):
-        """
-        Count rows in the bedsets table
-
-        :return int: number of rows in the bedsets table
-        """
-        return self._count_rows(table_name=BEDSET_TABLE)
-
-    def _drop_table(self, table_name):
-        """
-        Remove selected table from the database
-
-        :param str table_name: name of the table to remove
-        """
-        with self.db_cursor as cur:
-            statement = sql.SQL("DROP table IF EXISTS {}").format(
-                sql.Identifier(table_name))
-            cur.execute(statement)
-
-    def drop_bedfiles_table(self):
-        """
-        Remove bedfiles table from the database
-        """
-        self._drop_table(table_name=BED_TABLE)
-
-    def drop_bedsets_table(self):
-        """
-        Remove bedsets table from the database
-        """
-        self._drop_table(table_name=BEDSET_TABLE)
-
-    def drop_bedset_bedfiles_table(self):
-        """
-        Remove bedsets table from the database
-        """
-        self._drop_table(table_name=REL_TABLE)
-
-    def _create_table(self, table_name, columns):
-        """
-        Create a table
-
-        :param str table_name: name of the table to create
-        :param str | list[str] columns: columns definition list,
-            for instance: ['name VARCHAR(50) NOT NULL']
-        """
-        columns = _mk_list_of_str(columns)
-        with self.db_cursor as cur:
-            s = sql.SQL(f"CREATE TABLE {table_name} ({','.join(columns)})")
-            cur.execute(s)
-
-    def _get_columns_types(self, table_name):
-        """
-        Get types of the columns in the specified table
-
-        :param str table_name: name of the table to query
-        :return list[psycopg2.extras.DictRow]: column types
-        """
-        with self.db_cursor as cur:
-            cur.execute("SELECT column_name, data_type "
-                        "FROM information_schema.columns "
-                        "WHERE table_name = %s", (table_name, ))
-            return cur.fetchall()
-
-    def get_bedfiles_table_columns_types(self):
-        """
-        Get types of the columns in the bedfiles table
-
-        :return list[psycopg2.extras.DictRow]: column types
-        """
-        return self._get_columns_types(table_name=BED_TABLE)
-
-    def get_bedsets_table_columns_types(self):
-        """
-        Get types of the columns in the bedsets table
-
-        :return list[psycopg2.extras.DictRow]: column types
-        """
-        return self._get_columns_types(table_name=BEDSET_TABLE)
-
-    def get_bedset_bedfiles_table_columns_types(self):
-        """
-        Get types of the columns in the bedset_bedfiles table
-
-        :return list[psycopg2.extras.DictRow]: column types
-        """
-        return self._get_columns_types(table_name=REL_TABLE)
-
-    def create_bedfiles_table(self, columns):
-        """
-        Create a bedfiles table
-
-        :param str | list[str] columns: columns definition list,
-            for instance: ['name VARCHAR(50) NOT NULL']
-        """
-        self._create_table(table_name=BED_TABLE, columns=columns)
-
-    def create_bedsets_table(self, columns):
-        """
-        Create a bedsets table
-
-        :param str | list[str] columns: columns definition list,
-            for instance: ['name VARCHAR(50) NOT NULL']
-        """
-        self._create_table(table_name=BEDSET_TABLE, columns=columns)
-
-    def create_bedset_bedfiles_table(self):
+    def _create_bedset_bedfiles_table(self):
         """
         Create a bedsets table, id column is defined by default
         """
@@ -432,7 +161,24 @@ class BedBaseConf(yacman.YacAttMap):
                    f"{REL_BED_ID_KEY} INT NOT NULL",
                    f"FOREIGN KEY ({REL_BEDSET_ID_KEY}) REFERENCES {BEDSET_TABLE} (id)",
                    f"FOREIGN KEY ({REL_BED_ID_KEY}) REFERENCES {BED_TABLE} (id)"]
-        self._create_table(table_name=REL_TABLE, columns=columns)
+        self.bed._create_table(table_name=REL_TABLE, columns=columns)
+
+    def report_bedfile_for_bedset(self, bedset_id, bedfile_id):
+        """
+        Report a bedfile for bedset.
+
+        Inserts the ID pair into the relationship table, which allows to
+        manage many to many bedfile bedset relationships
+
+        :param int bedset_id: id of the bedset to report bedfile for
+        :param int bedfile_id: id of the bedfile to report
+        """
+        if not self.bed._check_table_exists(table_name=REL_TABLE):
+            self._create_bedset_bedfiles_table()
+        with self.bed.db_cursor as cur:
+            statement = f"INSERT INTO {REL_TABLE} " \
+                        f"({REL_BEDSET_ID_KEY},{REL_BED_ID_KEY}) VALUES (%s,%s)"
+            cur.execute(statement, (bedset_id, bedfile_id))
 
     def select_bedfiles_for_bedset(self, condition=None, condition_val=None,
                                    bedfile_col=None):
@@ -447,135 +193,17 @@ class BedBaseConf(yacman.YacAttMap):
         :return list[psycopg2.extras.DictRow]: matched bedfiles table contents
         """
         condition, condition_val = \
-            _preprocess_condition_pair(condition, condition_val)
-        columns = ["f." + c for c in _mk_list_of_str(bedfile_col or BED_COL_NAMES)]
+            pipestat.helpers.preprocess_condition_pair(condition, condition_val)
+        columns = ["f." + c for c in pipestat.helpers.mk_list_of_str(
+            bedfile_col or list(self.bed.schema.keys()))]
         columns = sql.SQL(',').join([sql.SQL(v) for v in columns])
         statement_str = \
             "SELECT {} FROM {} f INNER JOIN {} r ON r.bedfile_id = f.id INNER" \
             " JOIN {} s ON r.bedset_id = s.id WHERE s."
-        with self.db_cursor as cur:
+        with self.bed.db_cursor as cur:
             statement = sql.SQL(statement_str).format(
                 columns, sql.Identifier(BED_TABLE),
                 sql.Identifier(REL_TABLE), sql.Identifier(BEDSET_TABLE))
             statement += condition
             cur.execute(statement, condition_val)
             return cur.fetchall()
-
-    def _check_table_exists(self, table_name):
-        """
-        Check if the specified table exists
-
-        :param str table_name: table name to be checked
-        :return bool: whether the specified table exists
-        """
-        with self.db_cursor as cur:
-            cur.execute(
-                "SELECT EXISTS("
-                "SELECT * FROM information_schema.tables "
-                "WHERE table_name=%s)", (table_name,)
-            )
-            return cur.fetchone()[0]
-
-    def check_bedfiles_table_exists(self):
-        """
-        Check if the bedfiles table exists
-
-        :return bool: whether the bedfiles table exists
-        """
-        return self._check_table_exists(table_name=BED_TABLE)
-
-    def check_bedsets_table_exists(self):
-        """
-        Check if the bedsets table exists
-
-        :return bool: whether the bedsets table exists
-        """
-        return self._check_table_exists(table_name=BEDSET_TABLE)
-
-    def check_bedset_bedfiles_table_exists(self):
-        """
-        Check if the bedset_bedfiles table exists
-
-        :return bool: whether the bedset_bedfiles table exists
-        """
-        return self._check_table_exists(table_name=REL_TABLE)
-
-
-def _mk_list_of_str(x):
-    """
-    Make sure the input is a list of strings
-    :param str | list[str] | falsy x: input to covert
-    :return list[str]: converted input
-    :raise TypeError: if the argument cannot be converted
-    """
-    if not x or isinstance(x, list):
-        return x
-    if isinstance(x, str):
-        return [x]
-    raise TypeError(f"String or list of strings required as input. Got: "
-                    f"{x.__class__.__name__}")
-
-
-def get_bedbase_cfg(cfg=None):
-    """
-    Determine path to the bedbase configuration file
-
-    The path can be either explicitly provided
-    or read from a $BEDBASE environment variable
-
-    :param str cfg: path to the config file.
-        Optional, the $BEDBASE config env var will be used if not provided
-    :return str: configuration file path
-    """
-    selected_cfg = yacman.select_config(
-        config_filepath=cfg,
-        config_env_vars=CFG_ENV_VARS
-    )
-    if not selected_cfg:
-        raise BedBaseConnectionError(
-            f"You must provide a config file or set the "
-            f"{'or '.join(CFG_ENV_VARS)} environment variable")
-    return selected_cfg
-
-
-def _preprocess_condition_pair(condition, condition_val):
-    """
-    Preprocess query condition and values to ensure sanity and compatibility
-
-    :param str condition: condition string
-    :param tuple condition_val: values to populate condition string with
-    :return (psycopg2.sql.SQL, tuple): condition pair
-    """
-
-    def _check_semicolon(x):
-        """
-        recursively check for semicolons in an object
-
-        :param aby x: object to inspect
-        :raises ValueError: if semicolon detected
-        """
-        if isinstance(x, str):
-            assert ";" not in x, ValueError(
-                f"semicolons are not permitted in condition values: '{str(x)}'")
-        if isinstance(x, list):
-            list(map(lambda v: _check_semicolon(v), x))
-
-    if condition:
-        if not isinstance(condition, str):
-            raise TypeError("Condition has to be a string")
-        else:
-            _check_semicolon(condition)
-            placeholders = findall("%s", condition)
-            condition = sql.SQL(condition)
-        if not condition_val:
-            raise ValueError("condition provided but condition_val missing")
-        assert isinstance(condition_val, list), \
-            TypeError("condition_val has to be a list")
-        condition_val = tuple(condition_val)
-        assert len(placeholders) == len(condition_val), ValueError(
-            f"Number of condition ({len(condition_val)}) values not equal "
-            f"number of placeholders in: {condition}")
-    return condition, condition_val
-
-
-
